+++ src/davinci_crawling/task/api/serializers.py
@@ -47,6 +47,65 @@ class TaskSerializerV1(
                             "is_deleted", "status",
                             "times_performed", "type", "params_map",
                             "options_map")
+        extra_kwargs = {
+            'task_id': {
+                'help_text': 'the task id that is the unique partition key.'
+            },
+            'user': {
+                'help_text': 'The user that asked for the task, if it is an '
+                             'ondemand task.'
+            },
+            'created_at': {
+                'help_text': 'the date of the creation of the task.'
+            },
+            'updated_at': {
+                'help_text': 'the date that we last updated the task.'
+            },
+            'is_deleted': {
+                'help_text': 'controls if the data is deleted.'
+            },
+            'status': {
+                'help_text': """
+                 represents the actual status of the task, could be:
+                    - 0 (Created)
+                    - 1 (Queued)
+                    - 2 (In Progress)
+                    - 3 (Finished)
+                    - 4 (Faulty)
+                    - 5 (Unknown)
+                """
+            },
+            'kind': {
+                'help_text': 'the name of the crawler that will execute the '
+                             'task.'
+            },
+            'params': {
+                'help_text': 'the set of params used to execute the crawler '
+                             'command, this will be saved as Text.'
+            },
+            'params_map': {
+                'help_text': 'the exactly same content as `params` but saved '
+                             'on a way that we can search using solr '
+                             '(KeyEncodedMap).'
+            },
+            'options': {
+                'help_text': 'the set of options that is used to guide the '
+                             'crawler during the execution, this will be saved'
+                             ' as text.'
+            },
+            'options_map': {
+                'help_text': 'the exactly same content as `options` but saved'
+                             ' on a way that we can search using solr '
+                             '(KeyEncodedMap).'
+            },
+            'times_performed': {
+                'help_text': 'keep track on how many times the task was run.'
+            },
+            'type': {
+                'help_text': 'the type of the task, could be OnDemand(1) or '
+                             'Batch(2)'
+            },
+        }
 
 
 class TaskSearchSerializerV1(